requirements.txt
numpy == 1.18.5
scikit-learn == 0.23.2
scipy == 1.5.2.
tensorflow-gpu == 2.3.1
# keras==2.2.4
